refactor(cli): make find's --limit a usize like every other limit
FindArgs was the lone i64 limit, forcing .max(0) casts downstream; cmd_find
now takes usize and casts to i64 once at the SQL bind. MCP dispatch reads it
via uint like refs/grep/entries.

Deliberately NOT flattening the repeated --path/--limit args into one shared
clap struct: each command's help text describes its own scope semantics
("only symbols…" vs "only references…"), and a flatten would collapse
them into one generic line.

+++ src/commands/mcp_server.rs
@@ -315,7 +315,7 @@ fn mcp_call(
                 conn,
                 n,
                 opt("kind"),
-                int64("limit", defaults::FIND_LIMIT),
+                uint("limit", defaults::FIND_LIMIT),
                 opt("path"),
                 false,
             )?;
@@ -584,7 +584,7 @@ fn mcp_call(
                 conn,
                 sarg("name")?,
                 opt("kind"),
-                int64("limit", defaults::FIND_LIMIT),
+                uint("limit", defaults::FIND_LIMIT),
                 opt("path"),
                 true,
             ),
